splitting client and server
